Integrate Tests project into CMake build and adapt to Handle-based resource API
- Modified FANCY/CMakeLists.txt to unconditionally include Tests subdirectory
- Updated test files to use ResourceHandle<T> instead of SharedPtr<T>
- Fixed test constructors to work with Handle-based API and dereference handles via RenderCore::Get*()
- Updated Tests/CMakeLists.txt to explicitly list enabled test sources
- Removed Test_HazardTracking test (barrier tracking simplified in recent commits)
- Fixed Test_ImGui to use ImGuiCond_FirstUseEver instead of deprecated ImGuiSetCond_FirstUseEver
- Updated Test_Mipmapping to use static Assets class API instead of AssetManager
- Left Test_ModelViewer disabled pending more complex refactoring
- Fixed type conversions and warnings in Tests_main.cpp
- Tests executable now builds successfully and is included in PathTracer solution

+++ Tests/Tests_main.cpp
@@ -29,7 +29,7 @@ _Use_decl_annotations_ int WINAPI WinMain( HINSTANCE hInstance, HINSTANCE, LPSTR
 
   eastl::vector< eastl::string > commandLineArgStrings( numArgs );
   eastl::vector< const char * >  cStrings( numArgs );
-  for ( uint i = 0u; i < numArgs; ++i ) {
+  for ( int i = 0; i < numArgs; ++i ) {
     commandLineArgStrings[ i ] = StringUtil::ToNarrowString( commandLineArgs[ i ] );
     cStrings[ i ] = commandLineArgStrings[ i ].c_str();
   }
@@ -40,7 +40,7 @@ _Use_decl_annotations_ int WINAPI WinMain( HINSTANCE hInstance, HINSTANCE, LPSTR
   WindowParameters         windowParams;
   windowParams.myWidth = 1280;
   windowParams.myHeight = 720;
-  myTestApp.reset( new TestApplication( hInstance, cStrings.data(), cStrings.size(), "Tests", renderProperties, windowParams ) );
+  myTestApp.reset( new TestApplication( hInstance, cStrings.data(), static_cast<uint>( cStrings.size() ), "Tests", renderProperties, windowParams ) );
 
   MSG msg = { 0 };
   while ( true ) {
